requirements.txt
async-generator==1.10
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
census==0.8.17
certifi==2021.5.30
chardet==4.0.0
chromedriver==2.24.1
citipy==0.0.5
click==8.0.1
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
datareader==0.0.9
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
geojson==2.5.0
gmaps==0.9.0
gunicorn==20.1.0
idna==2.10
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
jupyter-core==4.6.3
jupyterlab-widgets==1.0.0
kdtree==0.16
kiwisolver==1.2.0
lxml==4.6.3
MarkupSafe==2.0.1
mistune==0.8.4
mysql==0.0.3
mysqlclient==2.0.3
opencv-python==4.5.2.54
packaging==20.4
pandas-datareader==0.9.0
pandocfilters==1.4.2
pickleshare==0.7.5
prometheus-client==0.8.0
psycopg2-binary==2.9.1
pymango==0.1.1
pymongo==3.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
selenium==3.141.0
simplejson==3.17.4
six==1.15.0
soupsieve==2.2.1
splinter==0.15.0
SQLAlchemy==1.3.17
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.26.5
webdriver-manager==3.4.2
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zope.interface==5.4.0
pickleshare==0.7.5